Refactored getData function Removed the getData function from router.js and moved to handler.js relates #23
